pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lowball"
version = "0.1.0"
description = "Benchmark for evaluating AI agents on car negotiation tasks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.0",
  "pyyaml>=6.0",
  "click>=8.0",
  "rich>=13.0",
  "openai>=1.0",
  "anthropic>=0.40",
  "pandas>=2.0",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
  "mypy>=1.10",
]

[project.scripts]
lowball = "lowball.cli:main"

[tool.ruff]
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
